refactor(actor): regroup ffi/inline*.rs under one ffi/inline/ module
Pure file move ahead of the per-instance inline-registry change (#1959):
`git mv` the suffix-sibling trio `ffi/inline.rs` / `ffi/inline_bundle.rs`
/ `ffi/inline_compose.rs` into one parent module `ffi/inline/{mod,bundle,
compose}.rs`, declare `mod bundle; pub mod compose;` in `inline/mod.rs`,
drop the old `pub mod inline_bundle; pub mod inline_compose;` decls, and
fix every intra-crate path and the `export!` macro string-literal paths
(`inline_compose::*` -> `inline::compose::*`, `inline_bundle::*` ->
`inline::bundle::*`). No behavior change.
